agent/requirements.txt
langchain
langchain-openai
openai
langchain-ollama
langchain-chroma
chromadb
langchain-community
pypdf
mcp[cli]
httpx
langchain-mcp-adapters
python-dotenv
langgraph
fastapi
uvicorn